don't abort when vbguest plugin not installed
When running vagrant up with --provider=lxc we shouldn't abort if the
virtualbox plugin is not installed. Unfortunately I don't know how to
check during Vagrant runtime which provider is being used so for now
this check is disabled.

Also sets memory usage to 2048M which is the same as the 2016 Magento
Grow Hypernode plan.