varnish caching is configurable in the config file
this adds a varnish enabled item to the config file. when varnish
enabled is false the nocache vcl is loaded. we need a vcl that passes
all requests to disable varnish instead of stopping the service because
otherwise a completely different nginx config would need to be placed.

also a workaround to ensure varnish is running so that vcls can be loaded during provisioning. in lxc vagrant boxes for some reason the varnish init script in /etc/init.d doesn't bring up the service on boot
todo: find out why varnish isn't always started on startup on lxc instances (on my machine?). varnish in lxc vagrants does work in the integration envirnoment.