copy default index.php if webdir is empty
Because there are two different webroots based on whether you are
booting a hypernode-vagrant with a magento 1 or magento 2 configuration
(by changing the local.yml) we would need to place the default index.php
in two places. Because symlinks on the host system are a pain because
people who use this box use many different filesystems which may or may
not support symlinks, let's just copy the index.php to the webdir if it
is empty during provisioning.