bundle vagrant-hypconfigmgmt gem with hypernode-vagrant
The basic hypernode-vagrant box can now be configured and booted by cloning the repo and running "vagrant up" three times and following the dialogue.
Vagrant needs to be started three consecutive times because it loads all configurations during initialization, not during runtime.
When we change the configuration in the Vagrant process we need to restart Vagrant for the changes to take effect.

1. Install the vagrant-hypconfigmgmt plugin
2. Configure the local.yml and install other required plugins
3. Boot the box

When the box is properly configured and all the required plugins is installed one vagrant up is sufficient.