do not compile Xdebug on Xenial
php5 is now php5.5 and some other things would be broken because now
both PHPs are installed in the same image. If anyone needs this I'd be
happy to script it (or create a PR yourself :)) but for now we'll
disable it to get the Xenial image out the door.